+++ resources/Views/auth/home.php
@@ -79,7 +79,8 @@ if (Auth::hasRole('Maintainer')) { ?>
 <div class="box">
     <strong><?= __('DeliveryNote') ?></strong><br>
     <a href="openVolumeDistributions"><?= __('OpenVolumeDistributions') ?></a><br>
-    <a href="showVolumeDistribution"><?= __('ShowVolumeDistribution') ?></a>
+    <a href="showVolumeDistribution"><?= __('ShowVolumeDistribution') ?></a><br>
+    <a href="amountDevelopmentStats"><?= __('AmountDevelopment') ?></a>
 </div>
 
 <div class="box">